成功完成了头部的拉伸效果..

+++ NTHeaderStretchingTableViewController/NTHeaderStretchingTableViewController/NTHeaderStretchingTableViewController.h
@@ -23,6 +23,6 @@
 /*
  * 导航栏背景图片名称
  */
-@property (nonatomic, assign) NSInteger headerImageHeight;
+@property (nonatomic, assign) int headerImageHeight;
 
 @end
